# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Marketplace YarmarOK 1.0.0",
    "description": "Web platform where vendors can come together to sell their products or services to customers.\n\nThis is Marketplace YarmarOK Server based on the OpenAPI 3.0 specification\n\nSome useful links:\n- [The YarmarokMarketplace repository](https://github.com/YarmarokMarketplace/marketplace)",
    "contact": {
      "name": "API developers",
      "email": "yarmarokmarketplace@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081/api",
      "description": "Development Server"
    },
    {
      "url": "https://yarmarok.onrender.com/api",
      "description": "Production Server"
    }
  ],
  "consumes": [
    "application/json",
    "multipart/form-data",
    "text/html"
  ],
  "produces": [
    "application/json",
    "multipart/form-data",
    "text/html"
  ],
...
  "paths": {
...
    "/user/{id}": {
      "delete": {
        "summary": "Delete user",
        "security": [{ "Bearer": [] }],
        "tags": [
          "User"
        ],
        "description": "Delete user account from the Database",
        "operationId": "removeUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user that needs to be deleted",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Delete the user with the specified ID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Notice not found"
          }
        }
      },
      "patch": {
        "summary": "Update user data",
        "security": [{ "Bearer": [] }],
        "tags": [
          "User"
        ],
        "description": "Update user info in the personal cabinet",
        "operationId": "updateUserData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user that needs to be updated",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Update the user with the specified ID"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/user/reviews/{id}": {
      "post": {
        "summary": "Add new notice",
        "tags": [
          "User"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Add new review",
        "operationId": "addReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of notice that needs to be assessed",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Update the user with the specified ID"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddReviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Olena"
          },
          "lastname": {
            "type": "string",
            "example": "Ivanova"
          },
          "patronymic": {
            "type": "string",
            "example": "Ivanivna"
          },
          "phone": {
            "type": "string",
            "example": "380961112233"
          },
          "avatarURL": {
            "type": "string",
            "example": "https://yarmarok-bucket.s3.amazonaws.com/1-kisb82VT8IOXQ3RbXa9.jpg"
          }
        }
      },
      "UpdateUserResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Backend-generated unique identifier",
            "example": "GGHJlhkj668__hjkhjkhh$%"
          },
          "email": {
            "type": "string",
            "example": "estasjuk@gmail.com"
          },
          "name": {
            "type": "string",
            "example": "Olena"
          },
          "lastname": {
            "type": "string",
            "example": "Ivanova"
          },
          "patronymic": {
            "type": "string",
            "example": "Ivanivna"
          },
          "phone": {
            "type": "string",
            "example": "380961112233"
          },
          "avatarURL": {
            "type": "string",
            "example": "https://yarmarok-bucket.s3.amazonaws.com/1-kisb82VT8IOXQ3RbXa9.jpg"
          }
        }
      },
...
      "AddReviewRequest": {
        "type": "object",
        "required": [
          "compliance",
          "delivery_speed",
          "communication"
        ],
        "properties": {
          "compliance": {
            "type": "number",
            "description": "Assessment of product compliance with declared characteristics",
            "enum": [1, 2, 3, 4, 5],
            "example": 5
          },
          "delivery_speed": {
            "type": "number",
            "description": "Assessment of product delivery speed",
            "enum": [1, 2, 3, 4, 5],
            "example": 5
          },
          "communication": {
            "type": "number",
            "description": "Assessment of communication with seller",
            "enum": [1, 2, 3, 4, 5],
            "example": 5
          },
          "comments": {
            "type": "string",
            "description": "Some additional information about product and seller assessment",
            "example": "Very good seller"
          }
        }
      },
      "AddReviewResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Backend-generated unique review identifier",
            "example": "GGHJlhkj668__hjkhjkhh$%"
          },
          "compliance": {
            "type": "number",
            "description": "Assessment of product compliance with declared characteristics",
            "enum": [1, 2, 3, 4, 5],
            "example": 5
          },
          "delivery_speed": {
            "type": "number",
            "description": "Assessment of product delivery speed",
            "enum": [1, 2, 3, 4, 5],
            "example": 5
          },
          "communication": {
            "type": "number",
            "description": "Assessment of communication with seller",
            "enum": [1, 2, 3, 4, 5],
            "example": 5
          },
          "comments": {
            "type": "string",
            "description": "Some additional information about product and seller assessment",
            "example": "Very good seller"
          },
          "product": {
            "type": "string",
            "description": "The product unique id",
            "example": "65001fef31a839a590ba9054"
          },
          "owner": {
            "type": "string",
            "description": "The user unique id who wrote this review",
            "example": "65001fef31a839a590ba9054"
          },
          "createdAt": {
            "type": "string",
            "description": "the creation time of review (backend-generated)",
            "example": "2023-05-15T10:51:30.531Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "the update time of notice (backend-generated),always equal to createdAt ",
            "example": "2023-05-15T10:51:30.531Z"
          }
        }
      }
    }
  },
...
}